feat: mainPin reset after form-submit\
