+++ js/show-modal.js
@@ -20,9 +20,8 @@ const newErrorModal = errorModalTemplate.cloneNode(true);
 
 const showModal = (modal) => {
   modalContainer.append(modal);
-  // modal.style.zIndex = '1000';
   modal.addEventListener('click', onClick(modal));
-  window.addEventListener('keydown', onModalEscKeydown(modal) );
+  // window.addEventListener('keydown', onModalEscKeydown(modal) );
 };
 
 
@@ -45,10 +44,10 @@ const onModalEscKeydown = (modal) => {
   }
 };
 
+
 const closeModal = (modal) => {
   modal.remove();
-  // modal.style.zIndex = '1000';
-  window.removeEventListener('keydown', onModalEscKeydown(modal));
+  // window.removeEventListener('keydown', onModalEscKeydown(modal));
   modal.removeEventListener('click', onClick(modal));
 };
